requirements.txt
certifi==2019.9.11
chardet==3.0.4
idna==2.8
lxml==4.4.1
Pillow==6.2.1
pkg-resources==0.0.0
requests==2.22.0
urllib3==1.25.6
